% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hillshader.R
\name{hillshader}
\alias{hillshader}
\title{Hillshader}
\usage{
hillshader(elevation, shader = "ray_shade", filename = NULL, ...)
}
\arguments{
\item{elevation}{Raster}

\item{shader}{Character. List of \code{rayshader} shader to sequentially apply. Default to \code{ray_shade}.}

\item{filename}{Character. If set, the result if written as a raster file.}

\item{...}{Additional parameters}
}
\description{
.
}
\author{
Pierre Roudier
}
