% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_to_raster.R
\name{matrix_to_raster}
\alias{matrix_to_raster}
\title{Matrix to Raster}
\usage{
matrix_to_raster(matrix, raster, crs = NULL)
}
\arguments{
\item{matrix}{The input matrix, typically the output of a \code{rayshader} operation}

\item{raster}{The original raster from which \code{matrix} is derived. Can be an \code{Extent} object.}

\item{crs}{If an \code{Extent} object is passed to the \code{raster} option, the corresponding coordinate reference system information.}
}
\description{
Turns a matrix into a Raster
}
\author{
Pierre Roudier
}
