% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maungawhau.R
\docType{data}
\name{maungawhau}
\alias{maungawhau}
\title{Elevation Raster for Maungawhau in T훮maki M훮kaurau/Auckland}
\format{
A \code{RasterLayer} with 87 rows, 61 columns, and 1 band with the elevation data at a 10-m resolution. The data is projected in New Zealand Map Grid (NZMG, EPSG:27200).
}
\source{
Elevation data from \code{datasets::volcano}, georeferencing adapted from \url{https://waterdata.usgs.gov/blog/inlmiscmaps}
}
\usage{
maungawhau
}
\description{
Elevation data as a raster for Maungawhau, a volcano located in T훮maki M훮kaurau/Auckland.
}
\keyword{datasets}
