% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maungawhau_hr.R
\docType{data}
\name{maungawhau_hr}
\alias{maungawhau_hr}
\title{LiDAR Elevation Raster for Maungawhau in T훮maki M훮kaurau/Auckland}
\format{
A \code{RasterLayer} with 860 rows, 600 columns, and 1 band with the elevation data at a 1 m resolution. The data is projected in New Zealand Map Grid (NZMG, EPSG:27200).
}
\source{
Elevation data from LINZ Data Service: \url{https://data.linz.govt.nz/layer/53405-auckland-lidar-1m-dem-2013/}
}
\usage{
maungawhau_hr
}
\description{
Elevation data as a raster for Maungawhau, a volcano located in T훮maki M훮kaurau/Auckland.
}
\keyword{datasets}
