% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_raster.R
\name{write_raster}
\alias{write_raster}
\title{Write hillshade to a file}
\usage{
write_raster(hillshade, elevation, filename, format, ...)
}
\arguments{
\item{hillshade}{A 2D matrix of shadow intensities.}

\item{elevation}{Original elevation raster.}

\item{filename}{Character.  Output filename.}

\item{format}{Character. Output file type. Passed to \code{raster::writeRaster}.}

\item{...}{Additional arguments passed to \code{raster::writeRaster}.}
}
\description{
Write an array from a hillshade procedure to a geospatial raster file.
}
\author{
Pierre Roudier
}
